getCross function for SMA algo added + lines when buy/sell
getCross function checks whereever the sma lines cross and produces a buy/sell annotation.

The reason some of the lines look off is how freq we get the data as the 5 minute interval can change a lot. THe 1 minute interval the lines add up.

The total is stored in a variable in the getCross function called 'totalMoney' but it is not shown on the homepage